+++ Cargo.toml
@@ -40,6 +40,7 @@ rayon = "1.3"
 serde = { version = "1.0.137", features = ["derive"] }
 wasmtime = { version = "3.0.0", default-features = false, features = ['cranelift'] }
 url = "2.0.0"
+pretty_assertions = "1.3.0"
 
 wasm-encoder = { version = "0.22.1", path = "crates/wasm-encoder"}
 wasm-compose = { version = "0.2.6", path = "crates/wasm-compose"}
@@ -48,7 +49,6 @@ wasm-mutate = { version = "0.2.17", path = "crates/wasm-mutate" }
 wasm-shrink = { version = "0.1.18", path = "crates/wasm-shrink" }
 wasm-smith = { version = "0.12.1", path = "crates/wasm-smith" }
 wasmparser = { version = "0.99.0", path = "crates/wasmparser" }
-wasmparser-dump = { version = "0.1.16", path = "crates/dump" }
 wasmprinter = { version = "0.2.49", path = "crates/wasmprinter" }
 wast = { version = "52.0.3", path = "crates/wast" }
 wat = { version = "1.0.57", path = "crates/wat" }
@@ -84,9 +84,6 @@ is_executable = { version = "1.0.1", optional = true }
 # Dependencies of `mutate`
 wasm-mutate = { workspace = true, features = ["clap"], optional = true }
 
-# Dependencies of `dump`
-wasmparser-dump = { workspace = true, optional = true }
-
 # Dependencies of `strip`
 wasm-encoder = { workspace = true, optional = true }
 regex = { version = "1.6.0", optional = true }
@@ -110,11 +107,11 @@ wasm-metadata = { workspace = true, features = ["clap"], optional = true }
 serde_json = "1.0"
 tempfile = "3.1"
 diff = "0.1"
-wasmparser-dump = { path = 'crates/dump' }
 wast = { path = 'crates/wast' }
+pretty_assertions = { workspace = true }
 
 [[test]]
-name = "dump"
+name = "cli"
 harness = false
 
 [[test]]
@@ -146,7 +143,7 @@ parse = []
 smith = ['wasm-smith', 'arbitrary', 'serde', 'serde_json']
 shrink = ['wasm-shrink', 'is_executable']
 mutate = ['wasm-mutate']
-dump = ['wasmparser-dump']
+dump = ['wasmparser']
 objdump = ['wasmparser']
 strip = ['wasm-encoder', 'wasmparser', 'regex']
 compose = ['wasm-compose']
