--- crates/dump/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "wasmparser-dump"
-version = "0.1.16"
-authors = ["The Wasmtime Project Developers"]
-edition.workspace = true
-license = "Apache-2.0 WITH LLVM-exception"
-repository = "https://github.com/bytecodealliance/wasm-tools/tree/main/crates/dump"
-description = "Utility to dump debug information about the wasm binary format"
-
-[dependencies]
-anyhow = { workspace = true }
-wasmparser = { workspace = true }
