+++ tests/cli/dump/alias.wat.stdout
