+++ tests/cli/dump/alias2.wat
@@ -1,3 +1,5 @@
+;; RUN: dump %
+
 (component
   (type $t (instance
     (export "a" (core module))
