+++ tests/cli/dump/alias2.wat.stdout
