+++ tests/cli/dump/blockty.wat.stdout
