+++ tests/cli/dump/bundled.wat.stdout
