+++ tests/cli/dump/component-expand-bundle.wat
@@ -1,3 +1,5 @@
+;; RUN: dump %
+
 (component
   (core module $m
     (func (export ""))
