+++ tests/cli/dump/component-expand-bundle.wat.stdout
