+++ tests/cli/dump/component-expand-bundle2.wat.stdout
