+++ tests/cli/dump/component-inline-export-import.wat
@@ -1,3 +1,5 @@
+;; RUN: dump %
+
 (component
   (component
     (export "")
