+++ tests/cli/dump/component-inline-export-import.wat.stdout
