+++ tests/cli/dump/component-inline-type.wat
@@ -1,3 +1,5 @@
+;; RUN: dump %
+
 (component
   (import "a" (component))
   (import "b" (core module))
