+++ tests/cli/dump/component-inline-type.wat.stdout
