+++ tests/cli/dump/component-instance-type.wat.stdout
