+++ tests/cli/dump/component-linking.wat.stdout
