+++ tests/cli/dump/component-outer-alias.wat.stdout
