+++ tests/cli/dump/instance-expand.wat.stdout
