+++ tests/cli/dump/instance-type.wat
@@ -1,3 +1,5 @@
+;; RUN: dump %
+
 (component
   (type (instance
     (export "" (func))
