+++ tests/cli/dump/instance-type.wat.stdout
