+++ tests/cli/dump/instance-type2.wat.stdout
