+++ tests/cli/dump/instantiate.wat.stdout
