+++ tests/cli/dump/instantiate2.wat
@@ -1,3 +1,5 @@
+;; RUN: dump %
+
 (component
  (import "a" (component $c (import "a" (func))))
  (instance (instantiate $c))
