+++ tests/cli/dump/instantiate2.wat.stdout
