+++ tests/cli/dump/module-types.wat.stdout
