+++ tests/cli/dump/names.wat
@@ -1,3 +1,5 @@
+;; RUN: dump %
+
 (module $foo
   (func $f  (param $x i32)
     (local $y f64)))
