+++ tests/cli/dump/names.wat.stdout
