+++ tests/cli/dump/nested-component.wat.stdout
