+++ tests/cli/dump/select.wat
@@ -1,3 +1,5 @@
+;; RUN: dump %
+
 (module
     (func
         select
