+++ tests/cli/dump/select.wat.stdout
