+++ tests/cli/dump/simple.wat.stdout
