+++ tests/cli/dump/try-delegate.wat
@@ -1,3 +1,5 @@
+;; RUN: dump %
+
 (module
   (func
     try $l
