+++ tests/cli/dump/try-delegate.wat.stdout
